+++ packages/website/src/pages/qrTest.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import QRCode from 'qrcode.react';
-import { useWeb3Context } from '../../../shared/contexts/web3Context';
+import { useWeb3Context } from '../contexts/web3Context';
 
 const generateURLWithSessionID = (sessionId) => {
     return `localhost:3000/attend?sessionId=${sessionId}`;
